refactor: 모바일 헤더를 단일 MobileHeader 컴포넌트로 통합
화면마다 제각각 인라인으로 구현된 '뒤로가기 + 타이틀' 모바일 sticky
헤더를 단일 규격 컴포넌트(app.component/layout/MobileHeader)로 통합한다.
타이틀 폰트가 화면마다 불일치(MobileStickyHeader heading2 vs 나머지
body1)하던 문제를 다수 관례인 body1(text-xl)/extrabold 로 통일한다.

통합 규격:
- 컨테이너: sticky top-0 z-30, h-14-safe + pt-safe-top(safe-area),
  border-b border-gray-100, bg-white/95 backdrop-blur, px-4, lg:hidden
- 뒤로가기: 32x32 rounded-lg, ArrowLeft 20px, hover:bg-gray-100
- 타이틀: DS Text body1/extrabold, tracking-[-0.3px], truncate
- 우측 액션 슬롯(right)·보조 문구(subtitle, 태블릿+) 옵션 지원

치환한 헤더 변종:
- 공용: MobileStickyHeader(삭제), SubPageShell 모바일 헤더
- 인라인: 챌린지 참여자/일지 리스트, 챌린지 생성/수정, 일지 상세/작성,
  회원 프로필, 설치 가이드, 일지 상세 스켈레톤
- LegalPageShell: 인라인 헤더 + LegalBackButton(삭제) 치환

design-system은 외부 published 패키지(@1d1s/design-system)로 소비하므로
동일 컴포넌트를 DS 레포에 반영하려면 별도 퍼블리시가 필요하다. 우선
클라 공용 레이어에서 통일한다.

동작 변경 없음(폰트 규격 통일 목적의 시각적 정정 제외).
